// backend/src/types/response/point/pointInsertResponse.ts
export type PointInsertResponse = {
  action_key: string;
  action_key_display_text: string;
  action_visible_order: number;
  action_icon: string;
  action_description: string;
  is_action_enabled: number;
  status: string;
  user_id: number;
  admin_ref: number;
  created_by: number;
  updated_by: number;
};